Data validation implemented

+++ StockMarketAPI/Controllers/StockController.cs
@@ -23,15 +23,23 @@ namespace StockMarketAPI.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
-            var stocks = await _stockRepo.GetAllAsync();
+			if (!ModelState.IsValid)
+			{
+				return BadRequest(ModelState);
+			}
+			var stocks = await _stockRepo.GetAllAsync();
             var stockDto = stocks.Select(s => s.ToStockDto());
             return Ok(stocks);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
-            var stock = await _stockRepo.GetByIdAsync(id);
+			if (!ModelState.IsValid)
+			{
+				return BadRequest(ModelState);
+			}
+			var stock = await _stockRepo.GetByIdAsync(id);
             if (stock == null)
             {
                 return NotFound();
@@ -42,16 +50,24 @@ namespace StockMarketAPI.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
-            var stockModel = stockDto.ToStockFromCreateDTO();
+			if (!ModelState.IsValid)
+			{
+				return BadRequest(ModelState);
+			}
+			var stockModel = stockDto.ToStockFromCreateDTO();
             await _stockRepo.CreateAsync(stockModel);
             return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
         {
-            var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
+			if (!ModelState.IsValid)
+			{
+				return BadRequest(ModelState);
+			}
+			var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
             if (stockModel == null)
             {
                 return NotFound();
@@ -60,10 +76,14 @@ namespace StockMarketAPI.Controllers
 		}
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
-            var stockModel = await _stockRepo.DeleteAsync(id);
+			if (!ModelState.IsValid)
+			{
+				return BadRequest(ModelState);
+			}
+			var stockModel = await _stockRepo.DeleteAsync(id);
             if (stockModel == null)
             {
                 return NotFound();
